fix: further folders restructure
