+++ packages/babel-plugin/src/utils/traverse-access-path/evaluate-path/namespace-import.tsx
@@ -1,8 +1,8 @@
 import * as t from '@babel/types';
 
-import type { Metadata } from '../../../../types';
-import { createResultPair } from '../../../create-result-pair';
-import { getDefaultExport, getNamedExport } from '../../../traversers';
+import type { Metadata } from '../../../types';
+import { createResultPair } from '../../create-result-pair';
+import { getDefaultExport, getNamedExport } from '../../traversers';
 
 export const evaluateNamespaceImportPath = (
   expression: t.Expression,
