+++ packages/babel-plugin/src/utils/traverse-access-path/evaluate-path/object.tsx
@@ -1,8 +1,8 @@
 import type * as t from '@babel/types';
 
-import type { Metadata } from '../../../../types';
-import { createResultPair } from '../../../create-result-pair';
-import { getObjectPropertyValue } from '../../../traversers';
+import type { Metadata } from '../../../types';
+import { createResultPair } from '../../create-result-pair';
+import { getObjectPropertyValue } from '../../traversers';
 
 export const evaluateObjectPath = (
   expression: t.ObjectExpression,
