+++ packages/babel-plugin/src/utils/traverse-access-path/index.tsx
@@ -1,11 +1,12 @@
 import type * as t from '@babel/types';
 
-import type { Metadata } from '../../../types';
-import { createResultPair } from '../../create-result-pair';
+import type { Metadata } from '../../types';
+import { createResultPair } from '../create-result-pair';
+import { traverseCallExpression } from '../traverse-expression';
 
 import { evaluatePath } from './evaluate-path';
 import { resolveExpressionInMember } from './resolve-expression';
-import type { TraverseHandlers } from './types';
+import type { TraverseHandlers, traverseMemberExpressionHandler } from './types';
 
 export const traverseMemberAccessPath = (
   expression: t.Expression,
@@ -13,8 +14,12 @@ export const traverseMemberAccessPath = (
   expressionName: string,
   accessPath: t.Identifier[],
   memberExpression: t.MemberExpression,
-  traversers: TraverseHandlers
+  traverseMemberExpression: traverseMemberExpressionHandler
 ): ReturnType<typeof createResultPair> => {
+  const traversers: TraverseHandlers = {
+    callExpression: traverseCallExpression,
+    memberExpression: traverseMemberExpression,
+  };
   const { value: resolvedExpression, meta: updatedMeta } = resolveExpressionInMember(
     expression,
     meta,
@@ -33,7 +38,7 @@ export const traverseMemberAccessPath = (
       pathName,
       accessPath.slice(1),
       memberExpression,
-      traversers
+      traverseMemberExpression
     );
   }
