+++ packages/babel-plugin/src/utils/traverse-access-path/resolve-expression/function-args.tsx
