+++ packages/babel-plugin/src/utils/traverse-access-path/resolve-expression/identifier.tsx
@@ -1,8 +1,8 @@
 import type * as t from '@babel/types';
 
-import type { Metadata } from '../../../../types';
-import { createResultPair } from '../../../create-result-pair';
-import { resolveBinding } from '../../../resolve-binding';
+import type { Metadata } from '../../../types';
+import { createResultPair } from '../../create-result-pair';
+import { resolveBinding } from '../../resolve-binding';
 
 export const evaluateIdentifier = (
   expression: t.Identifier,
